refactor(DUP): handle data in a separate module
Moves `handle_data` to its own module
